pyproject.toml
[tool.poetry]
name = "vectorapi"
version = "0.1.0"
description = ""
authors = ["Dennis Vink <pixelatedpoets@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
gunicorn = "^21.2.0"
faiss-cpu = "^1.7.3"
sentence-transformers = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
